+++ packages/crawling/package.json
@@ -8,10 +8,11 @@
   },
   "scripts": {
     "ky-open": "tsx src/findKYByOpen.ts",
-    "ky-youtube": "tsx src/crawling/crawlYoutube.ts",
-    "ky-verify": "tsx src/crawling/crawlYoutubeVerify.ts",
+    "ky-youtube": "tsx src/cron/crawlYoutube.ts",
+    "ky-verify": "tsx src/cron/crawlYoutubeVerify.ts",
     "ky-update": "pnpm run ky-youtube & pnpm run ky-verify",
-    "recent-tj": "tsx src/crawling/crawlRecentTJ.ts",
+    "recent-tj": "tsx src/cron/crawlRecentTJ.ts",
+    "tag-songs": "tsx src/cron/taggingSongs.ts",
     "lint": "eslint .",
     "test": "vitest run",
     "format": "prettier --write \"**/*.{ts,tsx,md}\""
